+++ compysition/util/__init__.py
@@ -1,3 +1,4 @@
+import warnings
 from contextlib import contextmanager
 import sys
 
@@ -9,6 +10,20 @@ def ignore(*exceptions):
 		yield
 	except exceptions:
 		pass
+
+_default_warning_filter = 'always'
+
+def set_warning_filter():
+    warnings.simplefilter(_default_warning_filter, DeprecationWarning)
+    warnings.simplefilter(_default_warning_filter, PendingDeprecationWarning)
+
+@contextmanager
+def suppress_deprecation():
+    warnings.simplefilter("ignore", DeprecationWarning)
+    warnings.simplefilter("ignore", PendingDeprecationWarning)
+    yield
+    warnings.simplefilter(_default_warning_filter, DeprecationWarning)
+    warnings.simplefilter(_default_warning_filter, PendingDeprecationWarning)
 
 def raise_(exception):
     raise exception